requirements.txt
absl-py==0.9.0
astunparse==1.6.3
atari-py==0.2.6
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
cloudpickle==1.3.0
decorator==4.4.2
dm-acme==0.1.7
dm-control==0.0.319497192
dm-env==1.2
#dm-reverb-nightly==0.1.0.dev20200616
dm-sonnet==2.0.0
dm-tree==0.1.5
dopamine-rl==3.1.2
future==0.18.2
gast==0.3.3
gin-config==0.3.0
glfw==1.11.2
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.30.0
gym==0.17.2
h5py==2.10.0
idna==2.10
Keras-Preprocessing==1.1.2
lxml==4.5.1
Markdown==3.2.2
numpy==1.19.0
oauthlib==3.1.0
opencv-python==4.3.0.36
opt-einsum==3.2.1
Pillow==7.2.0
#pkg-resources==0.0.0
portpicker==1.3.1
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyglet==1.5.0
PyOpenGL==3.1.5
pyparsing==2.4.7
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
six==1.15.0
tabulate==0.8.7
tb-nightly==2.3.0a20200706
tensorboard-plugin-wit==1.7.0
termcolor==1.1.0
tf-estimator-nightly==2.4.0.dev2020070701
#tf-nightly==2.3.0.dev20200616
tfp-nightly==0.11.0.dev20200707
trfl==1.1.0
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
